requirements.txt
langchain
faiss-cpu
unstructured
pypdf
tqdm
python-dotenv
requests
